+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Group /mskills skills by class
  - Fix very short visibility range of firework effects
  - Fix losing your broom if you die while riding
+ - Add "/mtrait invisible" for invisible NPCs
 
 ## 4.9.3